Add files via upload
Every year, an estimated 30-40% of food in the US goes in the trash. This amounts to around 60 million TONS per year. This map shows my local area, Pinellas and Hillsborough counties, and their food waste in tons. There are several composting programs, and this map details the net food waste, excluding waste that is composted.
Especially in Pinellas County, landfill space is scare. 80% of waste is incinerated and used to power homes. However, the remaining waste is exported to a landfill that is expected to fill within 80 years, with no potential new landfill locations to be found.
Throughout this spatial analysis, I measured the tons of non-composted food waste per zip code in my local area, encompassing Pinellas and Hillsborough Counties in Florida. I found that on average, 3746 tons of food waste per zip code went straight into the landfill. This figure is inferred to be annual as given in "per capita" metrics, although not explicitly mentioned by the data source (EPA, 2023). Depending on the zip code, anywhere from 1-12,000 tons of food waste went into the landfill.
There are many greener measures that can be taken, such as rerouting or repurposing waste, and decreasing food waste. Could composting more of our food waste help keep our landfills emptier?