Add files via upload
This map was created using data from the US Dept. of Ag.
The potential for farming on any given soil can be determined largely by soil composition, pH, and presence of urban structures. This series of maps details the small range of area that qualifies as "prime farmland" in the state, detailed in green. The red areas, which dominate the landscape, show land that is not prime farmland. While some of this is due to cities and construction, much of this is also due to soil quality.
The map shown in blue depicts areas that potentially could be farmed if not for frequent or present flooding.
It is clear that farming potential can be improved in our state- but how? There are several options available, but adding composted soil to areas that are otherwise unfit for farming can introduce valuable nutrients to raise crops successfully. This is a simple possible step we can take to help our state raise more produce independently!
Additionally, we can introduce farming into local areas on a smaller scale with composted materials and making the most of urban spaces. Making concrete or asphalt spaces greener can cool city temperatures (due to the urban heat effect) and provide fresh produce for locals.